+++ R/select_helpers.R
@@ -9,6 +9,9 @@
 #' - `all_interaction()` selects interaction terms from a regression model
 #' - `all_intercepts()` selects intercept terms from a regression model
 #' - `all_contrasts()` selects variables in regression model based on their type of contrast
+#' - `all_tests()` selects variables by the name of the test performed
+#' - `all_summary_cols()` selects columns from `tbl_summary`/`tbl_svysummary` object with summary statistics (i.e. "`stat_0`", "`stat_1`", "`stat_2`", etc.)
+#' @param tests character
 #' @name select_helpers
 #' @return A character vector of column names selected
 #' @examples
@@ -31,4 +34,24 @@ all_continuous2 <- function() {
   .generic_selector("variable", "var_type",
                     .data$var_type %in% "continuous2",
                     fun_name = "all_continuous")
+}
+
+#' @rdname select_helpers
+#' @export
+all_tests <- function(tests = NULL) {
+  if (is.null(tests) || !is.character(tests) || any(!tests %in% df_add_p_tests$test_name)) {
+    paste("The `tests=` argument must be one or more of the following:",
+          paste(shQuote(df_add_p_tests$test_name), collapse = ", ")) %>%
+      stop(call. = FALSE)
+  }
+
+  .generic_selector("variable", "test_name",
+                    .data$test_name %in% .env$tests,
+                    fun_name = "all_tests")
+}
+
+#' @rdname select_helpers
+#' @export
+all_summary_cols <- function() {
+  dplyr::matches("^stat_\\d+$") # finds stat_0, stat_1, stat_2, etc.
 }